# src/main/resources/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Esta é a documentação swagger da API para o TCC e-commerce de chinelos, ainda não está implementada e ainda faltam algumas actions, mas pretendemos continuar o projeto até o fim.",
    "version": "1.0.0",
    "title": "Slippers Swagger",
    "contact": {
      "email": "andrey_coradin@hotmail.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api",
...
  "paths": {
...
    "/endereco/buscarEnderecos": {
      "get": {
        "tags": [
          "Endereço"
        ],
        "summary": "Busque seus endereços",
        "description": "Buscar os endereços do usuário logado",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/EnderecoRetorno"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/endereco/inserirEndereco": {
      "post": {
        "tags": [
          "Endereço"
        ],
        "description": "Coloque um cep, e retornaremos o endereço completo.",
        "summary": "Inserção de endereços",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto para criação de endereço",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EnderecoInsert"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/EnderecoRetorno"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/chinelo/listaChinelos": {
      "get": {
        "tags": [
          "Chinelo"
        ],
        "summary": "Busca de chinelos",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nomeChinelo",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Nome do chinelo que você quer buscar. Caso não coloque nada, será retornada uma lista com todos os chinelos"
          },
          {
            "name": "pagina",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Página da paginação"
          },
          {
            "name": "qtdItensPagina",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Quantidade de itens por página."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Chinelo"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          }
        }
      }
    },
    "/chinelo/inserirChinelo": {
      "post": {
        "tags": [
          "Chinelo"
        ],
        "summary": "Inserção de chinelo",
        "description": "Apenas ADM pode usar esta função",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto para criação de chinelo",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChineloForm"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Chinelo"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/chinelo/atualizarChinelo/{id}": {
      "put": {
        "tags": [
          "Chinelo"
        ],
        "summary": "Atualização de chinelos",
        "description": "Apenas ADM pode usar esta função",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "id do chinelo a ser atualizado",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Objeto para atualização de chinelo. Tudo será alterado, por isso, preencha todos os campos",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChineloForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Chinelo"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/chinelo/deletarChinelo/{id}": {
      "delete": {
        "tags": [
          "Chinelo"
        ],
        "summary": "Deleção de chinelos",
        "description": "Apenas ADM tem acesso à esta função",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "id do chinelo a ser apagado",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Apagado com sucesso"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/carrinho/inserirChinelo": {
      "post": {
        "tags": [
          "Carrinho"
        ],
        "summary": "Inserção de chinelo no carrinho",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto para inserção de chinelo no carrinho",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CarrinhoForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Carrinho"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/carrinho/listarChinelos": {
      "get": {
        "tags": [
          "Carrinho"
        ],
        "summary": "Busca de Chinelos no carrinho",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Carrinho"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/carrinho/alterarQtdChineloCarrinho": {
      "put": {
        "tags": [
          "Carrinho"
        ],
        "summary": "Atualização de carrinho",
        "description": "Atualização do carrinho",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto para alterar chinelo do carrinho",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CarrinhoForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Carrinho"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/carrinho/deletarChinelo/{id}": {
      "delete": {
        "tags": [
          "Carrinho"
        ],
        "summary": "Remoção de chinelo do carrinho",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "id do chinelo a ser deletado",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Apagado com sucesso",
            "schema": {
              "$ref": "#/definitions/Carrinho"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira aqui o token resgatado.\nFormato: Bearer {token}\nExemplo: Bearer n8we9f9iom1f3289fj12f2mof123f9m2.f23"
    }
  },
  "definitions": {
    "400": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Mensagem de erro"
        }
      }
    },
...
    "EnderecoInsert": {
      "type": "object",
      "properties": {
        "cep": {
          "type": "string",
          "example": "12345678"
        },
        "complemento": {
          "type": "string"
        },
        "numero": {
          "type": "string"
        }
      }
    },
    "EnderecoRetorno": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "emailUsuarios": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "logradouro": {
          "type": "string"
        },
        "bairro": {
          "type": "string"
        },
        "cidade": {
          "type": "string"
        },
        "complemento": {
          "type": "string"
        },
        "numero": {
          "type": "string"
        },
        "uf": {
          "type": "string"
        }
      }
    },
    "Chinelo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "descricao": {
          "type": "string"
        },
        "valor": {
          "type": "number"
        },
        "urlImagem": {
          "type": "string"
        },
        "mediaNota": {
          "type": "integer"
        },
        "totalVendas": {
          "type": "integer"
        },
        "totalEstrelas": {
          "type": "integer"
        },
        "dataCriacao": {
          "type": "string",
          "example": "2020-05-05T12:35:47"
        },
        "tamanhosDTO": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Tamanho"
          }
        }
      }
    },
    "Tamanho": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "descricao": {
          "type": "string",
          "example": "40"
        }
      }
    },
    "ChineloForm": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string"
        },
        "descricao": {
          "type": "string"
        },
        "urlImagem": {
          "type": "string"
        },
        "valor": {
          "type": "number"
        },
        "idTamanhos": {
          "type": "array",
          "items": {
            "type": "integer",
            "example": 1
          }
        }
      }
    },
...
    "CarrinhoForm": {
      "type": "object",
      "properties": {
        "idChinelo": {
          "type": "integer"
        },
        "quantidade": {
          "type": "integer"
        }
      }
    },
    "Carrinho": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "emailUsuario": {
          "type": "string"
        },
        "valorTotalCarrinho": {
          "type": "number"
        },
        "quantidadeItensCarrinho": {
          "type": "integer"
        },
        "chinelosCarrinho": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Chinelo"
          }
        }
      }
    },
...
  }
}